refactor: Model, View 폴더 분리
